Ignore all the doc tests
This is unfortunate. Now that we're producing an rlib for the fuzzer, cargo
wants to test the library's documentation, and the documentation assumes the
library is called comrak. Even if the library were called comrak, the
markdown_to_html function the docs are using isn't being exported. This
seems like the most expeditious solution.