Fix test.sh's SPECS_ONLY variable (#89)
This was doing the exact opposite of what's intended. Now
it runs `cargo build` whenever SPECS_ONLY is not set.

I actually haven't been using this feature of the script
(obviously, since it's broken), so I'm conflicted about
whether to just remove it, but this is more correct,
and I _feel_ like I want to use this.