Validate media elements during rtjson postprocessing
This accomplishes several different things: first, it rejects media elements
that aren't "img", "gif" or "vid"; second, it eliminates nodes nested under
Image elements that the rtjson renderer can't render; finally, it rejects media
elements that don't occur in position expected from the RTE (in a top-level
paragraph).

This is vaguely similar to what the HTML renderer does for complex nested image
nodes (rendering their children as "plain" text).

It's kind of a distressing amount of logic to accomplish this, but I couldn't
come up with anything much better. I didn't want to put it directly in the
parser for the sake of isolating the reddit-specific logic (and because the
logic would have probably been even more complex there). Something similar to
this could be done in the rtjson renderer, but it seemed simpler to munge the
AST in postprocessing and just leave the renderer to do its thing.

Fixes CREATE-1650
Fixes CREATE-1444